Revert useAttributionListing.
